# docs/swagger.yaml
basePath: /api
definitions:
...
  dtos.ConfigResponse:
    properties:
      config_name:
        type: string
      program_id:
        type: integer
      value:
        type: string
    type: object
...
  dtos.CreateProgramRequest:
    properties:
      program_name:
        type: string
    type: object
...
  dtos.ProjectConfigResponse:
    properties:
      id:
        type: integer
      is_active:
        type: boolean
      program_id:
        type: integer
      title:
        type: string
    type: object
  dtos.ProjectConfigUpsertRequest:
    properties:
      id:
        type: integer
      is_active:
        type: boolean
      program_id:
        type: integer
      title:
        type: string
    type: object
...
  models.FileExtension:
    properties:
      extension_name:
        type: string
      id:
        type: integer
      mime_type:
        type: string
    type: object
...
  models.Program:
    properties:
      id:
        type: integer
      program_name:
        type: string
    type: object
...
  models.ProjectResourceConfig:
    properties:
      file_extension:
        $ref: '#/definitions/models.FileExtension'
      file_extension_id:
        type: integer
      icon_name:
        type: string
      id:
        type: integer
      max_file_size:
        type: integer
      program:
        $ref: '#/definitions/models.Program'
      program_id:
        type: integer
      resource_type:
        $ref: '#/definitions/models.ResourceType'
      resource_type_id:
        type: integer
      title:
        type: string
    type: object
...
  models.ResourceType:
    properties:
      id:
        type: integer
      type_name:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /v1/configs/program/{program_id}:
    get:
      description: Fetches all config for a given program
      parameters:
      - description: Program ID
        in: path
        name: program_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched config
          schema:
            items:
              $ref: '#/definitions/dtos.ConfigResponse'
            type: array
        "400":
          description: Invalid program ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Program not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get config by program ID
      tags:
      - Config
  /v1/programs:
    get:
      description: Retrieves all programs from the database
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched programs
          schema:
            items:
              $ref: '#/definitions/models.Program'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get All Programs
      tags:
      - Program
    post:
      consumes:
      - application/json
      description: Creates a new program in the database
      parameters:
      - description: Program creation details
        in: body
        name: program
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateProgramRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created program
          schema:
            $ref: '#/definitions/models.Program'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Create a New Program
      tags:
      - Program
  /v1/programs/update-name:
    put:
      consumes:
      - application/json
      description: Updates the name of an existing program
      parameters:
      - description: Program details with ID and updated name
        in: body
        name: program
        required: true
        schema:
          $ref: '#/definitions/models.Program'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated program name
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body or parameters
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Update Program Name
      tags:
      - Program
  /v1/projectConfigs:
    put:
      consumes:
      - application/json
      description: Update all project config if ID is provided, otherwise insert new
        project config
      parameters:
      - description: Configurations
        in: body
        name: configs
        required: true
        schema:
          items:
            $ref: '#/definitions/dtos.ProjectConfigUpsertRequest'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated config
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid program ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Program not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Upsert project config
      tags:
      - ProjectConfig
  /v1/projectConfigs/program/{program_id}:
    get:
      description: Fetches all config for a given program
      parameters:
      - description: Program ID
        in: path
        name: program_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched config
          schema:
            items:
              $ref: '#/definitions/dtos.ProjectConfigResponse'
            type: array
        "400":
          description: Invalid program ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Program not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get config by program ID
      tags:
      - ProjectConfig
  /v1/projectResourceConfigs:
    put:
      consumes:
      - application/json
      description: Insert or update project resource configurations. If an ID is provided,
        it updates the configuration; otherwise, it inserts a new configuration.
      parameters:
      - description: configuration to upsert
        in: body
        name: configs
        required: true
        schema:
          $ref: '#/definitions/models.ProjectResourceConfig'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully upsert configurations
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Upsert Project Resource Configurations
      tags:
      - ProjectResourceConfig
  /v1/projectResourceConfigs/program/{program_id}:
    get:
      description: Fetch all project resource configurations for a given program ID
      parameters:
      - description: Program ID
        in: path
        name: program_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched configurations
          schema:
            items:
              items:
                $ref: '#/definitions/models.ProjectResourceConfig'
              type: array
            type: array
        "400":
          description: Invalid program ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: No configurations found for the given program ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get Project Resource Config by Program ID
      tags:
      - ProjectResourceConfig
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"